add bootstrap and change line endings

+++ MyFirstWebsite/Controllers/MoviesController.cs
@@ -51,9 +51,13 @@ namespace MyFirstWebsite.Controllers
         //
         // GET: /Movies/Details/5
 
-        public ViewResult Details(int id)
+        public ActionResult Details(int id = 0)
         {
             Movie movie = db.Movies.Find(id);
+            if (movie == null)
+            {
+                return HttpNotFound();
+            }
             return View(movie);
         }
 
@@ -111,10 +115,14 @@ namespace MyFirstWebsite.Controllers
 
         //
         // GET: /Movies/Delete/5
- 
-        public ActionResult Delete(int id)
+
+        public ActionResult Delete(int id = 0)
         {
             Movie movie = db.Movies.Find(id);
+            if (movie == null)
+            {
+                return HttpNotFound();
+            }
             return View(movie);
         }
 
@@ -122,9 +130,13 @@ namespace MyFirstWebsite.Controllers
         // POST: /Movies/Delete/5
 
         [HttpPost, ActionName("Delete")]
-        public ActionResult DeleteConfirmed(int id)
-        {            
+        public ActionResult DeleteConfirmed(int id = 0)
+        {
             Movie movie = db.Movies.Find(id);
+            if (movie == null)
+            {
+                return HttpNotFound();
+            }
             db.Movies.Remove(movie);
             db.SaveChanges();
             return RedirectToAction("Index");
